set correct MJD output precision
for double we have 15 significant decimal digits. so we can have
10 decimal places for MJD.

so with 10 decimal places in MJD we can have 5 decimal places
for the second:

floor(MJD)*86400 = 8.64*10^-6 s = 1.0*10^-5 s

so correct output of date2MJD and MJD2date accordingly. and
also write correct offset/break information. a rounding
mechanism had to be implemented for second values, to round
to the 5th decimal place.